# [CANN] weight format to nz for Ascend310P3