**Why is this PR needed?**
When loading mat mul weights on Ascend 310P, convert them into an Ascend-friendly format to improve performance

**Performance Comparison**
Model：Qwen2-7B-Instruct-FP16

with nz feature:
>llama_perf_sampler_print:    sampling time =      23.57 ms /    71 runs   (    0.33 ms per token,  3011.92 tokens per second)
llama_perf_context_print:        load time =   16101.54 ms
llama_perf_context_print: prompt eval time =     237.11 ms /    21 tokens (   11.29 ms per token,    88.57 tokens per second)
llama_perf_context_print:        eval time =    7746.14 ms /    49 runs   (  158.08 ms per token,     6.33 tokens per second)
llama_perf_context_print:       total time =   10259.05 ms /    70 tokens
llama_perf_context_print:    graphs reused =          0

without nz feature:
> llama_perf_sampler_print:    sampling time =      22.02 ms /    71 runs   (    0.31 ms per token,  3224.34 tokens per second)
llama_perf_context_print:        load time =    7106.60 ms
llama_perf_context_print: prompt eval time =     441.01 ms /    21 tokens (   21.00 ms per token,    47.62 tokens per second)
llama_perf_context_print:        eval time =   18732.05 ms /    49 runs   (  382.29 ms per token,     2.62 tokens per second)
llama_perf_context_print:       total time =   32417.79 ms /    70 tokens
llama_perf_context_print:    graphs reused =          0