@hipudding 2025-06-28
Please paste MULMAT test case result and provide performance comparison data.